ml_online/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.9.0
stable-baselines3>=1.5.0
gymnasium>=0.26.0
requests>=2.25.0
websockets>=10.0
asyncio
uvloop
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
pyyaml>=5.4.0
tensorboard>=2.7.0